5 lesson

+++ src/lesson5/task1/Map.kt
@@ -313,9 +313,9 @@ fun findSumOfTwo(list: List<Int>, number: Int): Pair<Int, Int> {
  *   ) -> emptySet()
  */
 fun bagPacking(treasures: Map<String, Pair<Int, Int>>, capacity: Int): Set<String> {
-    var map = mapOf<String, Int>()
+    var map = mapOf<String, Double>()
     for ((key, value) in treasures) {
-        val cash = value.second / value.first
+        val cash = value.second.toDouble() / value.first.toDouble()
         map += Pair(key, cash)
     }
     val list= map.toList().sortedByDescending { (key, value) -> value }
